el total del order se calcula con los subtotales del detail

+++ src/main/java/org/example/apitiendaaa/domain/DTO/OrderOutDTO.java
@@ -10,7 +10,7 @@ import lombok.NoArgsConstructor;
 public class OrderOutDTO {
     private long id;
     private String status;
-    private int total;
+    private Float total;
     private int userId;
     private String address;
     private String creationDate;
